refactor: 로그인 관련 공통 로직 useAuth로 분리
- 공통로직은 Kakao 로그인에만 종속되는 개념이 아님
- 따라서 useAuth를 이용해 분리
